# docs/reference/openapi.kernel.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SoulOS Kernel",
    "version": "0.6.0"
  },
  "paths": {
...
    "/v1/avatars": {
      "post": {
        "tags": [
          "avatars"
        ],
        "summary": "Register Avatar",
        "operationId": "register_avatar_v1_avatars_post",
        "parameters": [
          {
            "name": "X-SoulOS-Account-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Soulos-Account-Id"
            }
          },
          {
            "name": "X-SoulOS-Gateway-Secret",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Soulos-Gateway-Secret"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/memory/ingest": {
      "post": {
        "tags": [
          "memory"
        ],
        "summary": "Ingest Memory",
        "operationId": "ingest_memory_memory_ingest_post",
        "parameters": [
          {
            "name": "X-SoulOS-Account-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Soulos-Account-Id"
            }
          },
          {
            "name": "X-SoulOS-Gateway-Secret",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Soulos-Gateway-Secret"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemoryIngest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/memory/retrieve": {
      "post": {
        "tags": [
          "memory"
        ],
        "summary": "Retrieve Memory",
        "operationId": "retrieve_memory_memory_retrieve_post",
        "parameters": [
          {
            "name": "X-SoulOS-Account-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Soulos-Account-Id"
            }
          },
          {
            "name": "X-SoulOS-Gateway-Secret",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Soulos-Gateway-Secret"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemoryRetrieve"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/memory/forget": {
      "post": {
        "tags": [
          "memory"
        ],
        "summary": "Forget Memory Route",
        "operationId": "forget_memory_route_memory_forget_post",
        "parameters": [
          {
            "name": "X-SoulOS-Account-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Soulos-Account-Id"
            }
          },
          {
            "name": "X-SoulOS-Gateway-Secret",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Soulos-Gateway-Secret"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemoryForget"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/hybrid/prepare": {
      "post": {
        "tags": [
          "hybrid"
        ],
        "summary": "Hybrid Prepare",
        "operationId": "hybrid_prepare_hybrid_prepare_post",
        "parameters": [
          {
            "name": "X-SoulOS-Account-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Soulos-Account-Id"
            }
          },
          {
            "name": "X-SoulOS-Gateway-Secret",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Soulos-Gateway-Secret"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HybridPrepareRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/hybrid/complete": {
      "post": {
        "tags": [
          "hybrid"
        ],
        "summary": "Hybrid Complete",
        "operationId": "hybrid_complete_hybrid_complete_post",
        "parameters": [
          {
            "name": "X-SoulOS-Account-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Soulos-Account-Id"
            }
          },
          {
            "name": "X-SoulOS-Gateway-Secret",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Soulos-Gateway-Secret"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HybridCompleteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Complete succeeded (sync reflect or reflect skipped)",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "202": {
            "description": "Accepted \u2014 async reflect"
          },
          "404": {
            "description": "TURN_SESSION_EXPIRED \u2014 turn session missing or TTL-purged"
          },
          "409": {
            "description": "TURN_STATE_STALE \u2014 expected_version mismatch (CAS lost race)"
          },
          "422": {
            "description": "TURN_CONTRACT_VIOLATION | TURN_REJECT_TOKEN | TURN_STEP_MISMATCH | request validation"
          }
        }
      }
    },
    "/chat/generate": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Chat Generate",
        "operationId": "chat_generate_chat_generate_post",
        "parameters": [
          {
            "name": "X-SoulOS-Account-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Soulos-Account-Id"
            }
          },
          {
            "name": "X-SoulOS-Gateway-Secret",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Soulos-Gateway-Secret"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/state/update": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Update State",
        "operationId": "update_state_state_update_post",
        "parameters": [
          {
            "name": "X-SoulOS-Account-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Soulos-Account-Id"
            }
          },
          {
            "name": "X-SoulOS-Gateway-Secret",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Soulos-Gateway-Secret"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ChatRequest": {
        "properties": {
          "bot_id": {
            "type": "string",
            "title": "Bot Id"
          },
          "message": {
            "type": "string",
            "maxLength": 32768,
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "bot_id",
          "message"
        ],
        "title": "ChatRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "HybridCompleteRequest": {
        "properties": {
          "bot_id": {
            "type": "string",
            "title": "Bot Id"
          },
          "summary": {
            "type": "string",
            "maxLength": 32768,
            "title": "Summary"
          },
          "user_message": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 32768
              },
              {
                "type": "null"
              }
            ],
            "title": "User Message"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "reflect": {
            "type": "boolean",
            "title": "Reflect",
            "default": true
          },
          "reflect_async": {
            "type": "boolean",
            "title": "Reflect Async",
            "default": false
          },
          "reflect_force": {
            "type": "boolean",
            "title": "Reflect Force",
            "default": false
          },
          "filled_slots": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filled Slots"
          },
          "intent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Intent"
          },
          "assistant_text": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 32768
              },
              {
                "type": "null"
              }
            ],
            "title": "Assistant Text"
          },
          "expected_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Version"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          },
          "advance": {
            "type": "boolean",
            "title": "Advance",
            "default": true
          },
          "expected_step": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Step"
          }
        },
        "type": "object",
        "required": [
          "bot_id",
          "summary"
        ],
        "title": "HybridCompleteRequest"
      },
      "HybridPrepareRequest": {
        "properties": {
          "bot_id": {
            "type": "string",
            "title": "Bot Id"
          },
          "query": {
            "type": "string",
            "maxLength": 32768,
            "title": "Query"
          },
          "top_k": {
            "type": "integer",
            "title": "Top K",
            "default": 5
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "include_shared_memory": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Include Shared Memory"
          }
        },
        "type": "object",
        "required": [
          "bot_id",
          "query"
        ],
        "title": "HybridPrepareRequest"
      },
...
      "MemoryForget": {
        "properties": {
          "bot_id": {
            "type": "string",
            "title": "Bot Id"
          },
          "content_match": {
            "type": "string",
            "maxLength": 32768,
            "title": "Content Match"
          }
        },
        "type": "object",
        "required": [
          "bot_id",
          "content_match"
        ],
        "title": "MemoryForget"
      },
      "MemoryIngest": {
        "properties": {
          "bot_id": {
            "type": "string",
            "title": "Bot Id"
          },
          "content": {
            "type": "string",
            "maxLength": 32768,
            "title": "Content"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          }
        },
        "type": "object",
        "required": [
          "bot_id",
          "content"
        ],
        "title": "MemoryIngest"
      },
...
      "MemoryRetrieve": {
        "properties": {
          "bot_id": {
            "type": "string",
            "title": "Bot Id"
          },
          "query": {
            "type": "string",
            "maxLength": 32768,
            "title": "Query"
          },
          "top_k": {
            "type": "integer",
            "title": "Top K",
            "default": 5
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "include_scores": {
            "type": "boolean",
            "title": "Include Scores",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "bot_id",
          "query"
        ],
        "title": "MemoryRetrieve"
      },
...
      "UpdateStateRequest": {
        "properties": {
          "bot_id": {
            "type": "string",
            "title": "Bot Id"
          },
          "new_msv": {
            "additionalProperties": true,
            "type": "object",
            "title": "New Msv"
          }
        },
        "type": "object",
        "required": [
          "bot_id",
          "new_msv"
        ],
        "title": "UpdateStateRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}